+++ pkg/core/src/tests/test-helpers/interfaces/IDivider.sol
@@ -2,7 +2,7 @@
 pragma solidity 0.8.11;
 
 abstract contract IDivider {
-    function series(address adapter, uint256 maturity)
+    function series(address adapter, uint48 maturity)
         external
         virtual
         returns (
@@ -12,32 +12,32 @@ abstract contract IDivider {
             address
         );
 
-    function initSeries(address adapter, uint256 maturity) external virtual returns (address zero, address claim);
+    function initSeries(address adapter, uint48 maturity) external virtual returns (address zero, address claim);
 
-    function settleSeries(address adapter, uint256 maturity) external virtual;
+    function settleSeries(address adapter, uint48 maturity) external virtual;
 
     function issue(
         address adapter,
-        uint256 maturity,
+        uint48 maturity,
         uint256 balance
     ) external virtual;
 
     function combine(
         address adapter,
-        uint256 maturity,
+        uint48 maturity,
         uint256 balance
     ) external virtual;
 
     function redeemZero(
         address adapter,
-        uint256 maturity,
+        uint48 maturity,
         uint256 balance
     ) external virtual;
 
     function collect(
         address usr,
         address adapter,
-        uint256 maturity,
+        uint48 maturity,
         uint256 balance
     ) external virtual returns (uint256 _collected);
 
@@ -45,7 +45,7 @@ abstract contract IDivider {
 
     function backfillScale(
         address adapter,
-        uint256 maturity,
+        uint48 maturity,
         uint256 scale,
         bytes[] memory backfills
     ) external virtual;
